+++ src/beta/openai/mod.rs
@@ -3,10 +3,14 @@
 *   Copyright © 2026 NatML Inc. All Rights Reserved.
 */
 
+mod chat;
+mod completions;
 mod embeddings;
 mod schema;
 mod utils;
 
+pub use chat::*;
+pub use completions::*;
 pub use embeddings::*;
 pub use schema::*;
 
@@ -15,6 +19,8 @@ use crate::services::{PredictionService, PredictorService};
 /// Experimental OpenAI client.
 #[derive(Clone)]
 pub struct OpenAIClient {
+    /// Chat service.
+    pub chat: ChatService,
     /// Embeddings service.
     pub embeddings: EmbeddingService,
 }
@@ -25,8 +31,8 @@ impl OpenAIClient {
         predictors: PredictorService,
         predictions: PredictionService
     ) -> Self {
-        Self {
-            embeddings: EmbeddingService::new(predictors, predictions),
-        }
+        let chat = ChatService::new(predictors.clone(), predictions.clone());
+        let embeddings = EmbeddingService::new(predictors, predictions);
+        Self { chat, embeddings }
     }
 }